student/requirements.txt
Django==2.2
gunicorn==20.0.4
djangorestframework==3.9.2